fix: quick fix for scryfall image cdn changes
Just using the dynamic urls instead. There's some additional cleanup here that could shift that logic into the UI which would eliminate the need to front load all that info. That would just rely on a "hasBack" field being added, or handling the 422's scryfall sends back in the UI.