feat: rudimentary stab at saving indexed set selections
I don't think I really want to get into mutating lists and tracking diffs on that, so this would be instead just tracking the number of each type of a card name in the list and storing that index for set selections.
eg. setSelections[name="fork"][index=2]

I want to figure out a cleaner pattern for this. Need to refresh myself on how javascript can handle defaulting nested objects.